# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project demo API documentation",
        "description": "A description for the API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/forecast/get": {
            "post": {
                "summary": "City forecast",
                "description": "Get city forecast by city name and country code",
                "operationId": "App\\Http\\Controllers\\ForeCast\\ForeCastGetController::__invoke",
                "requestBody": {
                    "description": "City and Country Code",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city",
                                    "country_code"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "GB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Wrong parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Sorry, wrong parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}